fix(defect): correct xml syntax errors
Because XML syntax uses some characters for tags and attributes it is not possible to directly use those characters inside XML tags or attribute values.

To include special characters inside XML files you must use the numeric character reference instead of that character. The numeric character reference must be UTF-8 because the supported encoding for XML files is defined in the prolog as encoding="UTF-8" and should not be changed.